moved state from apps js and got executePassbutton to half work
