+++ src/App.js
@@ -10,11 +10,9 @@ function App() {
     playerOneHand,
     playerTwoHand,
     playerThreeHand,
-    setPlayerOneHand,
-    setPlayerTwoHand,
-    setPlayerThreeHand,
+
     to,
-    setDeck,
+
     from,
     deck,
     setSelectedCard,
@@ -22,31 +20,6 @@ function App() {
 
     selectedCard,
   } = useGameContext();
-  function findCardIndex(value, suit, cards) {
-    return cards.findIndex((card) => card.value === value && card.suit === suit);
-  }
-
-  function passCard(card) {
-    const playerHands = [playerOneHand, playerTwoHand, playerThreeHand];
-    const playerHandSetFunctions = [setPlayerOneHand, setPlayerTwoHand, setPlayerThreeHand];
-
-    // arrays start at zero, but our players start at 1 :shrug:
-    const toHand = playerHands[to - 1] || deck;
-    const fromHand = playerHands[from - 1] || deck;
-
-    const toSetFunction = playerHandSetFunctions[to - 1] || setDeck;
-    const fromSetFunction = playerHandSetFunctions[from - 1] || setDeck;
-
-    const cardToMoveIndex = findCardIndex(card.value, card.suit, fromHand);
-    const [cardToMove] = fromHand.splice(cardToMoveIndex, 1);
-
-    toHand.push(cardToMove);
-
-    toSetFunction([...toHand]);
-    fromSetFunction([...fromHand]);
-
-    setSelectedCard(null);
-  }
 
   return (
     <div className="App">
@@ -63,18 +36,7 @@ function App() {
           player={'deck'}
         />
       </section>
-      <section>
-        {selectedCard && (
-          <ExecutePassButton
-            passCard={passCard}
-            setFrom={setFrom}
-            from={from}
-            to={to}
-            selectedCard={selectedCard}
-            setSelectedCard={setSelectedCard}
-          />
-        )}
-      </section>
+      <section>{selectedCard && <ExecutePassButton />}</section>
     </div>
   );
 }
