+++ src/components/ExecutePassButton.js
@@ -1,14 +1,50 @@
 import React from 'react';
+import { useGameContext } from '../context/GameContext';
 import Card from './Card';
 
-export default function ExecutePassButton({
-  passCard,
-  setFrom,
-  from,
-  to,
-  selectedCard,
-  setSelectedCard,
-}) {
+export default function ExecutePassButton() {
+  const {
+    selectedCard,
+    setSelectedCard,
+    from,
+    setFrom,
+    to,
+    playerOneHand,
+    playerTwoHand,
+    playerThreeHand,
+    setPlayerOneHand,
+    setPlayerTwoHand,
+    setPlayerThreeHand,
+    deck,
+    setDeck,
+  } = useGameContext();
+
+  function findCardIndex(value, suit, cards) {
+    return cards.findIndex((card) => card.value === value && card.suit === suit);
+  }
+
+  function passCard(card) {
+    const playerHands = [playerOneHand, playerTwoHand, playerThreeHand];
+    const playerHandSetFunctions = [setPlayerOneHand, setPlayerTwoHand, setPlayerThreeHand];
+
+    // arrays start at zero, but our players start at 1 :shrug:
+    const toHand = playerHands[to - 1] || deck;
+    const fromHand = playerHands[from - 1] || deck;
+
+    const toSetFunction = playerHandSetFunctions[to - 1] || setDeck;
+    const fromSetFunction = playerHandSetFunctions[from - 1] || setDeck;
+
+    const cardToMoveIndex = findCardIndex(card.value, card.suit, fromHand);
+    const [cardToMove] = fromHand.splice(cardToMoveIndex, 1);
+
+    toHand.push(cardToMove);
+
+    toSetFunction([...toHand]);
+    fromSetFunction([...fromHand]);
+
+    setSelectedCard(null);
+  }
+
   return (
     <div className="execute-button" onClick={() => passCard(selectedCard)}>
       Pass <Card card={selectedCard} setSelectedCard={setSelectedCard} setFrom={setFrom} /> from{' '}
