+++ src/components/Player.js
@@ -1,21 +1,16 @@
 import React from 'react';
+import { useGameContext } from '../context/GameContext';
 import CardList from './CardList';
 
-export default function Player({
-  player,
-  setTo,
-  hand,
-  setSelectedCard,
-
-  to,
-}) {
+export default function Player({ player, hand }) {
+  const { to, setTo } = useGameContext();
   return (
     <div
       className={`player ${to === player ? 'selected-player' : ''}`}
       onClick={() => setTo(player)}
     >
       <p>Player {player}</p>
-      <CardList player={player} cards={hand} setSelectedCard={setSelectedCard} />
+      <CardList player={player} cards={hand} />
     </div>
   );
 }
